nati_x86_64_defconfig: Disable the cpufreq driver to improve RT performance
The default configuration on NI Linux RT is optimized for performance
rather than power savings. As such the ability of the cpufreq driver
to dynamically scale the CPU frequency is not used.

cpufreq adds code in the scheduler hot paths and results in increased
latency for real-time threads. Disable it.

Natinst-AZDO-ID: 59877
Natinst-AZDO-ID: 1214772
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Alex Stewart <alex.stewart@ni.com>