artemis: add artemis into hardware model compatibility list
As artmemis has a design of combining zynq processor with sd card (ext4)
instead of NAND drive, a unique compatible string can be added to notify
the kernel if the target is artemis via the device tree. This string will
be used as a flag to allow artemis configuration scripts to run (eg.
postint for systemlink installation).

This commit adds "ni,artemis" into the model compatibility list.

Signed-off-by: wkoe <wilson.koe@ni.com>