arm: dts: ni-vb: declare partitions in device tree
We've been sending NAND partition information from u-boot to the kernel
via cmdline:

   mtdparts=xilinx_nand:128k(fsbl)ro,512k(u-boot)ro,40M(boot-config),-(root)

Shipping versions of the bootloader have a bug that restricts the kernel
cmdline to 256 bytes, so that space is at a premium. There exists a
perfectly good mechanism for declaring partitions via the device tree
instead-- by doing it this way we can free up 74 characters of cmdline
space.

Verified that /proc/mtd lists the same partitions on a VB-8034 before
this change as it does after the change with the `mtdparts` variable
removed from the cmdline.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>