i2c-i801.c: Skip SPD initialization on cRIO-903x
The cRIO-903x architecture is sort of special, we do not connect
a dimm to the SPD because we use flash instead, but the SPD is still
present because the processor expects it.  However, enumerating and
registering the SPD i2c bus is causing interrupt storms, so for now
we skip the registration on all 903x devices.

Fixes: 01590f3 ("i2c: i801: Instantiate SPD EEPROMs automatically")
Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Natinst-AZDO-ID: 1573148
(cherry picked from commit 4de019e72df37f49e08f352a835cf5779d57fac2)
[cvadrevu: fix trivial conflict with d08cac0a6378c ("i2c: i801: reword according to newest specification")]
[cvadrevu: fix minor conflict with 80e56b86b59e7 ("i2c: i801: Simplify class-based client device instantiation")]
Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@emerson.com>
[gratian: fix minor conflict with 4d6d35d3417d ("i2c: smbus: introduce Write Disable-aware SPD instantiating functions")]
Signed-off-by: Gratian Crisan <gratian.crisan@emerson.com>