sched: Consider task_struct::saved_state in wait_task_inactive().
Ptrace is using wait_task_inactive() to wait for the tracee to reach a
certain task state. On PREEMPT_RT that state may be stored in
task_struct::saved_state while the tracee blocks on a sleeping lock.

In that case wait_task_inactive() should wait until the requested state
is in task_struct::__state and the task idle.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>