nati_x86_64_defconfig: disable CONFIG_RCU_NOCB_CPU_CB_BOOST
The CONFIG_RCU_NOCB_CPU_CB_BOOST option can be used to invoke offloaded RCU
callbacks as SCHED_FIFO to avoid starvation by heavy SCHED_OTHER background
load.

However this can impact other latency-sensitive tasks and requires moving
them to other cores or ensuring higher priority as the help states:

  "Therefore, when enabling this option, it is your responsibility to
  ensure that latency-sensitive tasks either run with higher priority or
  run on some other CPU."

Since we are not in control of the final user application and we optimize
our platform for low latency - disable it. This matches the behavior of
previous NILRT kernels.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>