8250: make disabling CONFIG_SERIAL_8250_NI16550 work correctly
The legacy PNP IDs for NI 16550 devices (which is how 8250_pnp enumerates
them) are not gated by CONFIG_SERIAL_8250_NI16550, so if you turn that option
off, you end up with NI 16550s in an unfortunate partially-enumerated state
because 8250_pnp finds them, but doesn't initialize them properly.

This change puts those IDs behind ifdefs so that it becomes possible to
disable the existing NI 16550 driver.

Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>