serial: 8250: Mark txvr_ops as a deprecated interface
We added an interface in 8e9e45935119 ("8250: Allow client drivers to
control tranceivers") for RS-485 transceiver control, but this never
went upstream and is also not a requirement for a client driver to
implement RS-485 transceiver control.

However, this out-of-tree interface got exposed to the NI-Serial drivers,
and they were using it for their driver for the NI 987x devices, so we
can't (yet) drop it outright.

Mark this interface as deprecated and emit a warning when registering a
client driver that intends to use it.

Upstream-Status: Inappropriate [NI-specific]
Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>