Revert "usb: dwc3: gadget: Add 1ms delay after end transfer command without IOC"
This reverts commit 1450c82a16bba232dca0ff667b17c0970ddecb85.

That commit caused usblan to not work reliably between some cRIOs (903x,
9041) and windows host.
The issue happens only sometimes and when it happens, windows host would
report the network adapter as unconnected. The only way to reliably
recover from this was to reconnect usb cable or disable and enable
windows network adapter. Rebooting the cRIO would also sometimes fix the
issue.

Natinst-AzDo-ID: 2419039

Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@ni.com>