sdhci.c: Update tuning function to return errors
The sdhci_execute_tuning function was not actually returning the result
of the underlying tuning function. This caused multiple retuning
attempts for DDR50 cards, which cannot always be tuned, that led to an
I/O error. This change updates sdhci_execute_tuning to return the error
code which causes tuning to be disabled after the first attempt.

Signed-off-by: Erick Shepherd <erick.shepherd@ni.com>