net: dsa: mv88e6xxx: embed rx timestamps into packets
Marvell's "Generation 3" and later PTP switches have support for embedding
the switch timestamp counter into incoming packets, which eliminates
timestamp overwrites/discards due to not being able to read timestamps out
over SMI in time when receiving many packets at once.

On applicable hardware, we configure the switch to embed the timestamp
into the "reserved2" field. (Older switches have to read out the timestamp
counter the current way.)

We still need to queue the packet for a worker because we need to take
the reg lock, but we can simplify down to only needing a single rx queue
instead of splitting based on PTP packet type.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>