net: dsa: mv88e6xxx: add 1588-related addresses to ATU as management frames
The mv88e6xxx driver uses the hardware's trunking capabilities in order to
offload link aggregation. When packets ingress a trunked port, the DSA tag
that is added has a trunk ID rather than a port ID.

This is undesirable for when doing 1588 timestamping, because timestamping
is port-centric, but when we receive a 1588 packet in a LAG, the DSA
infrastructure doesn't know which port it came from and just delivers it
to the bond/team interface.

This behavior can be avoided by marking frames as being "management" in the
ATU; therefore, define a set of "extra management DAs" to be configured
with priority for every VLAN. This set currently includes destination
addresses relevant for L2 and L4 1588.

This is not needed for 802.1AS, because 802.1AS packets use a destination
address of 01:80:c2:00:00:0e, which is already covered by the rsvd2cpu
functionality.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>